refactor(search-router): split indexed-dirs + locate concerns into sibling modules
`src/nexus/server/api/v2/routers/search.py` was 2312 lines — above
the 2000-line health-gate. Two prior split-tracked-as-follow-up PRs
(#3818 postgres_profile_store, #4069 async_files) show the pattern
of exempting-instead-of-splitting slowly bloats source files past
useful review size.

This PR splits the two most encapsulated concern-sections into
sibling modules that the parent `search.py` includes at module-load
via FastAPI's `router.include_router`:

- `_search_indexed_dirs.py` (~500 lines) — the 5 endpoints for
  per-directory semantic-index scoping (Issue #3698):
  register / unregister / list indexed dirs, set indexing mode,
  purge unscoped. Includes the shared `_require_admin_or_path_write`
  policy gate that only these endpoints use.
- `_search_locate.py` (~150 lines) — the `/locate` endpoint
  (Issue #3725, BM25-lite path+title search).

Sub-modules are underscore-prefixed so no external caller is
tempted to import them directly — they only exist to attach
endpoints to the parent `router`. They import shared helpers
(`_get_search_daemon`) from `search.py` at their own module top;
`search.py`'s `from . import _search_*` calls stay at the very
END of the file so the parent's namespace is fully populated
before the sub-modules' imports resolve. Moving them higher would
circular-fault.

Query-section endpoints (query / batch / health / stats / parked)
and `_handle_federated_search` / `_handle_single_zone_search`
STAY in `search.py`. Nine integration tests use
`monkeypatch.setattr(search_mod, "_handle_federated_search", ...)`
which only works if the target lives in the parent module — keeping
that section co-located preserves those tests. The remaining
grep/glob/index/refresh/expand endpoints are candidates for a
future PR if search.py drifts back over 2000 lines.

Result: search.py 2312 → 1721 lines. All URL paths byte-identical
(`/api/v2/search/*`). No wire changes.
